Event-driven state overhaul part 5 - Final * Add create card * Add update card
